shop-backend/requirements.txt
alipay-sdk-python==3.6.72
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.6.0
awesome-slugify==1.6.5
black==21.6b0
click==8.0.1
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-versatileimagefield==2.2
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.6.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==9.0.0
promise==2.3
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycryptodome==3.14.1
PyJWT==1.7.1
python-magic==0.4.25
pytz==2021.1
regex==2021.7.1
rsa==4.8
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
Unidecode==0.4.21
zipp==3.5.0
